add compton config file
note that for some reason, disabling fading entirely (i.e. setting fading=false;)
seems to cause the desktop wallpaper to flash briefly when switching workspaces,
which is headache-inducing when using dark terminals and a light wallpaper.

the (temporary) fix seems to be this: set fade-in-step and fade-out-step = 1 to
make the workspaces instantly "appear" and "disappear." this seems to make
compton happy.